feat: 사이드바 통일
