+++ app/staff/class/layout.tsx
@@ -1,9 +1,9 @@
-import PageSidebar from "@/components/common/PageSidebar";
+import StaffSidebar from "@/components/staff/StaffSidebar";
 
 export default function ClassLayout({ children }: { children: React.ReactNode }) {
   return (
     <div className="flex min-h-screen">
-      <PageSidebar />
+      <StaffSidebar />
       <main className="flex-1">{children}</main>
     </div>
   );
